fix selection of blast hits, split annotationswidget.h to .h and .cpp
